predict welding fuel consumption
